Refactored redundant code
I gave spanish, french, and english their own blocks of code to make it
more readable and easy to follow.

There was also no need for the "fudge" variable since we could just
multiply base by the exact type of fudge. In the last else bracket
fudge is not used at all so I remove it entirely

The calculation for sugar is the same across all different recipe types
so I moved it below the if/else statement as it can be calculated at the
end once the correct "amt" is determined.